// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** Int or string */
  ActionCount: { input: any; output: any; }
  /** AreaOfEffectType ("SPHERE", "CUBE", "CYLINDER", "LINE", "CONE") or list of AreaOfEffectTypes */
  AreaOfEffectTypeFilter: { input: any; output: any; }
  /** Float, list of floats, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  FloatFilter: { input: any; output: any; }
  /** Int, list of ints, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  IntFilter: { input: any; output: any; }
  /** LanguageScript ("COMMON", "ELVISH", "DWARVISH", "INFERNAL", "DRACONIC", "CELESTIAL") or list of LanguageScripts */
  LanguageScriptFilter: { input: any; output: any; }
  /** MonsterSubtype ("ANY_RACE", "HUMAN", "DWARF", "ELF", "GOBLINOID", "MERFOLK", "SHAPECHANGER", "DEMON", "DEVIL", "ORC", "SAHUAGIN", "TITAN", "KOBOLD", "GNOLL", "GRIMLOCK", "LIZARDFOLK", "GNOME") or list of MonsterSubtypes */
  MonsterSubtypeFilter: { input: any; output: any; }
  /** MonsterTypes ("BEAST", "MONSTROSITY", "DRAGON", "HUMANOID", "UNDEAD", "FIEND", "CELESTIAL", "CONSTRUCT", "GIANT", "ELEMENTAL", "FEY", "ABERRATION", "OOZE", "SWARM", "PLANT") or list of MonsterTypes */
  MonsterTypeFilter: { input: any; output: any; }
  /** ProficiencyType ("WEAPONS", "ARTISANS_TOOLS", "SKILLS", "ARMOR", "MUSICAL_INSTRUMENTS", "SAVING_THROWS", "OTHER", "GAMING_SETS", "VEHICLES") or list of ProficiencyTypes */
  ProficiencyTypeFilter: { input: any; output: any; }
  /** Size ("TINY", "SMALL", "MEDIUM", "LARGE", "HUGE", "GARGANTUAN") or list of sizes */
  SizeFilter: { input: any; output: any; }
  /** SpellAttackType ("MELEE", "RANGED") or list of SpellAttackTypes */
  SpellAttackTypeFilter: { input: any; output: any; }
  /** String or list of strings */
  StringFilter: { input: any; output: any; }
};
// ...
export type Cost = {
  __typename?: 'Cost';
  quantity: Scalars['Int']['output'];
  unit: Currency;
};

export type CountedReferenceOption = {
  __typename?: 'CountedReferenceOption';
  count: Scalars['Int']['output'];
  of: IEquipment;
  option_type: Scalars['String']['output'];
  prerequisites?: Maybe<Array<ProficiencyPrerequisite>>;
};

export enum Currency {
  Cp = 'CP',
  Gp = 'GP',
  Sp = 'SP'
}

export type Damage = {
  __typename?: 'Damage';
  damage_dice: Scalars['String']['output'];
  damage_type: DamageType;
};
// ...
export type DamageType = {
  __typename?: 'DamageType';
  desc: Array<Scalars['String']['output']>;
  index: Scalars['String']['output'];
  name: Scalars['String']['output'];
};
// ...
export type EquipmentCategory = {
  __typename?: 'EquipmentCategory';
  equipment: Array<IEquipmentBase>;
  index: Scalars['String']['output'];
  name: Scalars['String']['output'];
};
// ...
export type IEquipment = {
  cost: Cost;
  desc?: Maybe<Array<Scalars['String']['output']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String']['output'];
  name: Scalars['String']['output'];
  weight?: Maybe<Scalars['Float']['output']>;
};

export type IEquipmentBase = {
  desc?: Maybe<Array<Scalars['String']['output']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String']['output'];
  name: Scalars['String']['output'];
};
// ...
export type Range = {
  __typename?: 'Range';
  long?: Maybe<Scalars['Int']['output']>;
  normal: Scalars['Int']['output'];
};
// ...
export type Weapon = IEquipment & IEquipmentBase & {
  __typename?: 'Weapon';
  category_range: EquipmentCategory;
  cost: Cost;
  damage?: Maybe<Damage>;
  desc?: Maybe<Array<Scalars['String']['output']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String']['output'];
  name: Scalars['String']['output'];
  properties: Array<WeaponProperty>;
  range: Range;
  special?: Maybe<Array<Scalars['String']['output']>>;
  throw_range?: Maybe<Range>;
  two_handed_damage?: Maybe<Damage>;
  weapon_category: EquipmentCategory;
  weapon_range: WeaponRange;
  weight?: Maybe<Scalars['Float']['output']>;
};


export type WeaponPropertiesArgs = {
  name?: InputMaybe<Scalars['String']['input']>;
};

export type WeaponProperty = {
  __typename?: 'WeaponProperty';
  desc: Array<Scalars['String']['output']>;
  index: Scalars['String']['output'];
  name: Scalars['String']['output'];
};

export enum WeaponRange {
  Melee = 'MELEE',
  Ranged = 'RANGED'
}